+++ resources/views/users/terms.blade.php
@@ -1,0 +1,76 @@
+@extends('layouts.app')
+
+@section('title', config('app.description'))
+
+@section('content')
+<div class="row terms">
+    <div class="col-md-8 offset-md-2">
+        <h1 class="page-header">JobsToMail Terms of Service and Privacy Policy</h1>
+
+        <h3>1. Terms</h3>
+
+        <p>By accessing the website at&nbsp;<a href="http://www.jobstomail.com">http://www.jobstomail.com</a>, you are agreeing to be bound by these terms of service, all applicable laws and regulations, and agree that you are responsible for compliance with any applicable local laws. If you do not agree with any of these terms, you are prohibited from using or accessing this site. The materials contained in this website are protected by applicable copyright and trademark law.</p>
+
+        <h3>2. Use License</h3>
+
+        <ol type="a">
+            <li>
+                Permission is granted to download the source materials (information or software) on JobsToMail's website for any use. This project is open source and licensed under the <a href="https://github.com/jobapis/jobs-to-mail/blob/master/license.md" target="_blank">Apache 2.0 license.</a> You may do anything with this site except the following:
+
+                <ol type="i">
+                    <li>attempt to access or download the raw user data from the database;</li>
+                    <li>remove any copyright or other proprietary notations from the materials; or</li>
+                    <li>transfer the materials to another person or "mirror" the materials on any other server without attribution.</li>
+                </ol>
+            </li>
+            <li>This license shall automatically terminate if you violate any of these restrictions and may be terminated by JobsToMail at any time. Upon terminating your viewing of these materials or upon the termination of this license, you must destroy any downloaded materials in your possession whether in electronic or printed format.</li>
+        </ol>
+
+        <h3>3. Disclaimer</h3>
+
+        <ol type="a">
+            <li>The materials on JobsToMail's website are provided on an 'as is' basis. JobsToMail makes no warranties, expressed or implied, and hereby disclaims and negates all other warranties including, without limitation, implied warranties or conditions of merchantability, fitness for a particular purpose, or non-infringement of intellectual property or other violation of rights.</li>
+            <li>Further, JobsToMail does not warrant or make any representations concerning the accuracy, likely results, or reliability of the use of the materials on its website or otherwise relating to such materials or on any sites linked to this site.</li>
+        </ol>
+
+        <h3>4. Limitations</h3>
+
+        <p>In no event shall JobsToMail or its suppliers be liable for any damages (including, without limitation, damages for loss of data or profit, or due to business interruption) arising out of the use or inability to use the materials on JobsToMail's website, even if JobsToMail or a JobsToMail authorized representative has been notified orally or in writing of the possibility of such damage. Because some jurisdictions do not allow limitations on implied warranties, or limitations of liability for consequential or incidental damages, these limitations may not apply to you.</p>
+
+        <h3>5. Accuracy of materials</h3>
+
+        <p>The materials appearing on JobsToMail's website could include technical, typographical, or photographic errors. JobsToMail does not warrant that any of the materials on its website are accurate, complete or current. JobsToMail may make changes to the materials contained on its website at any time without notice. However JobsToMail does not make any commitment to update the materials.</p>
+
+        <h3>6. Links</h3>
+
+        <p>JobsToMail has not reviewed all of the sites linked to its website and is not responsible for the contents of any such linked site. The inclusion of any link does not imply endorsement by JobsToMail of the site. Use of any such linked website is at the user's own risk.</p>
+
+        <h3>7. Modifications</h3>
+
+        <p>JobsToMail may revise these terms of service for its website at any time without notice. By using this website you are agreeing to be bound by the then current version of these terms of service.</p>
+
+        <h3>8. Governing Law</h3>
+
+        <p>These terms and conditions are governed by and construed in accordance with the laws of Chicago, IL and you irrevocably submit to the exclusive jurisdiction of the courts in that State or location.</p>
+
+        <h2>Privacy Policy</h2>
+
+        <p>Your privacy is important to us.</p>
+
+        <p>It is JobsToMail's policy to respect your privacy regarding any information we may collect while operating our website. Accordingly, we have developed this privacy policy in order for you to understand how we collect, use, communicate, disclose and otherwise make use of personal information. We have outlined our privacy policy below.</p>
+
+        <ul>
+            <li>We will collect personal information by lawful and fair means and, where appropriate, with the knowledge or consent of the individual concerned.</li>
+            <li>Before or at the time of collecting personal information, we will identify the purposes for which information is being collected.</li>
+            <li>We will collect and use personal information solely for fulfilling those purposes specified by us and for other ancillary purposes, unless we obtain the consent of the individual concerned or as required by law.</li>
+            <li>Personal data should be relevant to the purposes for which it is to be used, and, to the extent necessary for those purposes, should be accurate, complete, and up-to-date.</li>
+            <li>We will protect personal information by using reasonable security safeguards against loss or theft, as well as unauthorized access, disclosure, copying, use or modification.</li>
+            <li>We will make readily available to customers information about our policies and practices relating to the management of personal information.</li>
+            <li>We will only retain personal information for as long as necessary for the fulfilment of those purposes.</li>
+        </ul>
+
+        <p>We are committed to conducting our business in accordance with these principles in order to ensure that the confidentiality of personal information is protected and maintained. JobsToMail may change this privacy policy from time to time at JobsToMail's sole discretion.</p>
+        <small>Genereated by <a href="http://getterms.io/" target="_blank">GetTerms.io</a>.</small>
+    </div>
+</div>
+@endsection