+++ changelog.md
@@ -1,19 +1,20 @@
 # Changelog
 All Notable changes to `jobs-to-mail` will be documented in this file.
 
-## 0.2.0
+## 0.2.0 - 2016-09-29
 
 ### Added
 - Keyword and location parameters to jobs email.
 - Heroku setup instructions.
+- Terms and privacy policy.
 
 ## Fixed
-- Old references were loading bootstrap 3 instead of 4
+- Old references were loading bootstrap 3 instead of 4.
 
 ## Security
 - Use `random_bytes()` instead of `openssl_random_pseudo_bytes()` for secure token generation.
 
-## 0.1.0 - 2016-25-11
+## 0.1.0 - 2016-09-25
 
 ### Added
 - Initial pre-release.